Add new CLI to package.
